add pause menu, game over menu. add functionality to speed power up. Organize code with Game Manager, Game UI Controller.

+++ Assets/Scripts/PowerUpShieldBehaviour.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 
 public class PowerUpShieldBehaviour : MonoBehaviour, ICollectable
 {
-    [SerializeField] private int _shieldAmount = 1;
-    [SerializeField] private float _shieldDuration = 5;
+    [SerializeField] private float _shieldAmount = 1f;
+    [SerializeField] private float _shieldDuration = 5f;
 
-    public static event Action<int,float> OnPowerUpShieldCollected;
+    public static event Action<float,float> OnPowerUpShieldCollected;
     public void Collect()
     {
         OnPowerUpShieldCollected?.Invoke(_shieldAmount, _shieldDuration);
